+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
-- Limudim support for Parsha, DafYomiBavli
+- Limudim support for Parsha, DafYomiBavli, DafYomiYerushalmi
 - Kviah function for JewishDate
 
 ### Fixed